training tuned poly10 agent

+++ envs/random_polygon_env.py
@@ -239,7 +239,7 @@ class RandomPolygonEnv(gym.Env):
         return action_mask
 
     def _get_progress(self):
-        return np.array([self.num_steps / self.max_steps])
+        return np.array([self.num_steps / self.max_steps], dtype=np.float32)
 
     def _get_blank_obs(self):
         features = np.zeros((self.template_size, self.num_features), dtype=np.float32)
